Refactor caching to simplify tests
