+++ pydiscamb/discamb_wrapper.py
@@ -20,67 +20,14 @@ class FCalcMethod(Enum):
     TAAM = 1
 
 
-class DiscambWrapper(PythonInterface):
-
-    __cache = {}
+class DiscambWrapper_Uncached(PythonInterface):
 
     def __init__(self, xrs: structure, method: FCalcMethod = None, **kwargs):
-        """
-        Initialize a wrapper object for structure factor calculations using DiSCaMB.
-        Pass an xray structure and either a FCalcMethod,
-        or kwargs which are passed to  C++: :code:`discamb::SfCalculator::create`,
-        found in discamb/Scattering/SfCalculator.h
-
-        Notes
-        -----
-        The :code:`method` parameter is used for lookup of sensible default
-        values to be sent to :code:`discamb::SfCalculator::create`.
-        Any kwargs present will override the defaults.
-        The defaults are e.g. inferring the scattering table from `xrs`,
-        and using the default TAAM databank.
-        See :code:`DiscambWrapper._prepare_calculator_params` for details.
-
-        Parameters
-        ----------
-        xrs : structure
-            Object defining the symmetry, atom positions, ADPs, occupancy ect.
-        method : FCalcMethod, optional
-            Which structure factor model to use, by default FCalcMethod.IAM
-        """
-        if self.__check_cache(xrs, method, kwargs) is not None:
-            self.update_structure(xrs)
-            return
         calculator_params = self._prepare_calculator_params(xrs, method, kwargs)
         super().__init__(xrs, calculator_params)
 
         self._scatterer_flags = xrs.scatterer_flags()
         self._atomstr = _concat_scatterer_labels(xrs)
-
-        self.__cache[self.__get_cache_key(xrs, method, kwargs)] = self
-
-    def __new__(cls, xrs: structure, method: FCalcMethod = None, **kwargs):
-        cache = cls.__check_cache(xrs, method, kwargs)
-        if cache is None:
-            return super().__new__(cls)
-        return cache
-
-    @classmethod
-    def __check_cache(cls, xrs: structure, method: FCalcMethod, kwargs: Dict[str, str]):
-        key = cls.__get_cache_key(xrs, method, kwargs)
-        return cls.__cache.get(key)
-
-    @classmethod
-    def __get_cache_key(
-        cls, xrs: structure, method: FCalcMethod, kwargs: Dict[str, str]
-    ):
-        atomstr = _concat_scatterer_labels(xrs)
-        params = cls._prepare_calculator_params(xrs, method, kwargs)
-        key = (atomstr, *sorted(params.items()))
-        return key
-
-    @classmethod
-    def __clear_cache(cls):
-        cls.__cache = {}
 
     @staticmethod
     def _prepare_calculator_params(
@@ -115,7 +62,7 @@ class DiscambWrapper(PythonInterface):
     @classmethod
     def from_file(
         cls, filepath: Union[str, Path], method: FCalcMethod = FCalcMethod.IAM, **kwargs
-    ) -> "DiscambWrapper":
+    ) -> "DiscambWrapper_Uncached":
         """Read a structure from a file
 
         Parameters
@@ -164,7 +111,7 @@ class DiscambWrapper(PythonInterface):
     @classmethod
     def from_pdb_code(
         cls, pdb_code: str, method: FCalcMethod = FCalcMethod.IAM, **kwargs
-    ) -> "DiscambWrapper":
+    ) -> "DiscambWrapper_Uncached":
         """Download structure from rcsb.org
 
         Parameters
@@ -202,7 +149,7 @@ class DiscambWrapper(PythonInterface):
         return cls._from_pdb_str(pdb_str, method, **kwargs)
 
     @classmethod
-    def _from_pdb_str(cls, pdb_str: str, method, **kwargs) -> "DiscambWrapper":
+    def _from_pdb_str(cls, pdb_str: str, method, **kwargs) -> "DiscambWrapper_Uncached":
         import mmtbx.model
         import iotbx.pdb
         from libtbx.utils import null_out
@@ -213,7 +160,7 @@ class DiscambWrapper(PythonInterface):
 
     ## Annotations
     @overload
-    def f_calc(self, miller_array: None) -> flex.complex_double:
+    def f_calc(self, miller_set: None) -> flex.complex_double:
         ...
 
     @overload
@@ -221,7 +168,7 @@ class DiscambWrapper(PythonInterface):
         ...
 
     @overload
-    def f_calc(self, miller_array: miller.set) -> miller.array:
+    def f_calc(self, miller_set: miller.set) -> miller.array:
         ...
 
     @overload
@@ -308,10 +255,65 @@ class DiscambWrapper(PythonInterface):
         return out
 
 
+class DiscambWrapper(DiscambWrapper_Uncached):
+
+    __cache = {}
+
+    def __init__(self, xrs: structure, method: FCalcMethod = None, **kwargs):
+        """
+        Initialize a wrapper object for structure factor calculations using DiSCaMB.
+        Pass an xray structure and either a FCalcMethod,
+        or kwargs which are passed to  C++: :code:`discamb::SfCalculator::create`,
+        found in discamb/Scattering/SfCalculator.h
+
+        Notes
+        -----
+        The :code:`method` parameter is used for lookup of sensible default
+        values to be sent to :code:`discamb::SfCalculator::create`.
+        Any kwargs present will override the defaults.
+        The defaults are e.g. inferring the scattering table from `xrs`,
+        and using the default TAAM databank.
+        See :code:`DiscambWrapper_Uncached._prepare_calculator_params` for details.
+
+        Parameters
+        ----------
+        xrs : structure
+            Object defining the symmetry, atom positions, ADPs, occupancy ect.
+        method : FCalcMethod, optional
+            Which structure factor model to use, by default FCalcMethod.IAM
+        """
+        if self.__check_cache(xrs, method, kwargs) is not None:
+            self.update_structure(xrs)
+            return
+        
+        super().__init__(xrs, method, **kwargs)
+
+        self.__cache[self.__get_cache_key(xrs, method, kwargs)] = self
+
+    def __new__(cls, xrs: structure, method: FCalcMethod = None, **kwargs):
+        cache = cls.__check_cache(xrs, method, kwargs)
+        if cache is None:
+            return super().__new__(cls)
+        return cache
+
+    @classmethod
+    def __check_cache(cls, xrs: structure, method: FCalcMethod, kwargs: Dict[str, str]):
+        key = cls.__get_cache_key(xrs, method, kwargs)
+        return cls.__cache.get(key)
+
+    @classmethod
+    def __get_cache_key(
+        cls, xrs: structure, method: FCalcMethod, kwargs: Dict[str, str]
+    ):
+        atomstr = _concat_scatterer_labels(xrs)
+        params = cls._prepare_calculator_params(xrs, method, kwargs)
+        key = (atomstr, *sorted(params.items()))
+        return key
+
 def _calculate_structure_factors(
     xrs: structure, d_min: float, method: FCalcMethod
 ) -> List[complex]:
-    w = DiscambWrapper(xrs, method)
+    w = DiscambWrapper_Uncached(xrs, method)
     w.set_d_min(d_min)
     return w.f_calc()
